+++ Code/J-Nistler/django/PokeBusters/pokedex/templates/pokedex/index.html
@@ -20,11 +20,11 @@
       <table class="table table-striped table-hover list" width="100%">
         <thead class="thead-dark">
           <tr>
-            <th class="non-numeric" scope="col">Title</th>
-            <th scope="col">Year</th>
-            <th scope="col">Service</th>
-            <th scope="col">License Expires</th>
-            <th scope="col">URL</th>
+            <th scope="col">Number</th>
+            <th scope="col">Image</th>
+            <th scope="col">Name</th>
+            <th scope="col">Height</th>
+            <th scope="col">Weight</th>
           </tr>
         </thead>
         <tbody></tbody>
@@ -34,8 +34,8 @@
 </div>
 <script src="{% static 'pokedex/list.js' %}"></script>
 <script>
-  let list = JSON.parse("{{pokemons}}");
-  console.log(list);
-//   renderList(list);
+    // renderList(list);
+    let list = "{{ pokemons|safe }}";
+    console.log(typeof list);
 </script>
 {% endblock %}
