+++ Code/J-Nistler/html_css/README.md
@@ -1,0 +1,21 @@
+# J-Nistler HTML/CSS Code Directory
+
+## About this Directory
+
+The purpose of this directory is to host the HTML and CSS files for J-Nistler for the CodeBuster's class.
+
+## About the Files
+
+_This directory contains the HTML/CSS labs:_
+
+- `labs` - Individual challenges
+
+## Progress
+
+### Labs
+
+- [x] 01 Bio
+- [x] 02 Blog
+- [ ] 03 Company Page
+- [ ] 04 Burrito Order Form
+- [ ] 05 Personal Portfolio
