+++ Code/J-Nistler/html_css/lab03_company_page/equipment.html
@@ -1,0 +1,163 @@
+<!DOCTYPE html>
+<html lang="en">
+
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <title>Rankdog & Nisbone Brewery</title>
+  <link href="https://fonts.googleapis.com/css2?family=Ubuntu&display=swap" rel="stylesheet">
+  <link href="./css/styles.css" rel="stylesheet">
+
+  <!-- Bootstrap v4.4.1 CSS stylesheet -->
+  <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css"
+    integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
+</head>
+
+<body>
+  <header class="sticky-top">
+    <!-- Header Start -->
+    <nav class="navbar navbar-expand-sm navbar-light bg-light">
+      <div class="container">
+        <!-- Logo -->
+        <div class="logo">
+          <a href="./index.html"><img src="./images/nomad_weblogo.svg" alt=""></a>
+        </div>
+        <!-- Main-menu -->
+        <a class="navbar-brand" href="./index.html">Rankdog & Nisbone Brewery</a>
+        <button class="navbar-toggler navbar-toggler-right" type="button" data-toggle="collapse"
+          data-target="#navbarResponsive" aria-controls="navbarResponsive" aria-expanded="false"
+          aria-label="Toggle navigation">
+          <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarResponsive">
+          <ul class="navbar-nav ml-auto">
+            <li class="nav-item">
+              <a class="nav-link" href="./about.html">About</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" href="./equipment.html">Equipment</a>
+            </li>
+            <li class="nav-item">
+              <a class="nav-link" href="./material.html">Ingredients</a>
+            </li>
+          </ul>
+        </div>
+      </div>
+    </nav>
+    <!-- Header End -->
+  </header>
+  <main>
+    <!-- Jumbotron -->
+    <div class="jumbotron container-fluid" id="jumbo-equip">
+      <div class="container">
+        <h1 class="display-5">Equipment</h1>
+        <p>The machines we use everyday</p>
+        <p><a class="btn btn-info btn-lg" href="#equip-upgrades" role="button">Check them out</a></p>
+      </div>
+    </div>
+    <!-- Jumbotron End -->
+    <div class="container">
+      <h2>BrewMaster 5000</h2>
+      <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Veritatis laborum maxime facilis voluptate voluptatibus deserunt quos beatae dignissimos, dicta iure ratione repudiandae architecto aut voluptas! Quaerat fugiat in necessitatibus voluptas.
+      </p>
+      <h4>Current Specifications:</h4>
+      <table>
+        <thead>
+          <tr>
+            <th>Specifications</th>
+            <th>BrewMaster 5000</th>
+          </tr>
+        </thead>
+        <tbody>
+          <tr>
+            <td>Brewing Technology</td>
+            <td>FDM</td>
+          </tr>
+          <tr>
+            <td>Volume</td>
+            <td>235 x 235 x 250</td>
+          </tr>
+          <tr>
+            <td>Chassis Size</td>
+            <td>440 x 410 x 465</td>
+          </tr>
+          <tr>
+            <td>Stock Brewing Assembly</td>
+            <td>MK10 Assembly</td>
+          </tr>
+          <tr>
+            <td>User Interface</td>
+            <td>LCD Display with Rotary Encoder</td>
+          </tr>
+          <tr>
+            <td>Chassis Materials</td>
+            <td>Black Anodised Aluminium / V-Slot Extrusion</td>
+          </tr>
+        </tbody>
+      </table>
+    </div>
+    <hr class="featurette-divider">
+    <!--Circles-->
+  </main>
+
+  <footer>
+    <!-- Footer Start-->
+    <div class="footer-area">
+      <div class="container">
+        <div class="footer-top footer-padding">
+          <div class="row justify-content-center">
+            <div class="col-lg-12 text-center">
+              <!-- No Links listed as this website is not actually operating as a business-->
+              <p>
+                <a href="#">
+                  <svg aria-labelledby="simpleicons-twitter-icon" role="img" viewBox="0 0 24 24"
+                    xmlns="http://www.w3.org/2000/svg">
+                    <title id="simpleicons-twitter-icon">Twitter icon</title>
+                    <path
+                      d="M23.954 4.569c-.885.389-1.83.654-2.825.775 1.014-.611 1.794-1.574 2.163-2.723-.951.555-2.005.959-3.127 1.184-.896-.959-2.173-1.559-3.591-1.559-2.717 0-4.92 2.203-4.92 4.917 0 .39.045.765.127 1.124C7.691 8.094 4.066 6.13 1.64 3.161c-.427.722-.666 1.561-.666 2.475 0 1.71.87 3.213 2.188 4.096-.807-.026-1.566-.248-2.228-.616v.061c0 2.385 1.693 4.374 3.946 4.827-.413.111-.849.171-1.296.171-.314 0-.615-.03-.916-.086.631 1.953 2.445 3.377 4.604 3.417-1.68 1.319-3.809 2.105-6.102 2.105-.39 0-.779-.023-1.17-.067 2.189 1.394 4.768 2.209 7.557 2.209 9.054 0 13.999-7.496 13.999-13.986 0-.209 0-.42-.015-.63.961-.689 1.8-1.56 2.46-2.548l-.047-.02z" />
+                  </svg>
+                </a>
+                <a href="#">
+                  <svg aria-labelledby="simpleicons-facebook-icon" role="img" viewBox="0 0 24 24"
+                    xmlns="http://www.w3.org/2000/svg">
+                    <title id="simpleicons-facebook-icon">Facebook icon</title>
+                    <path
+                      d="M22.676 0H1.324C.593 0 0 .593 0 1.324v21.352C0 23.408.593 24 1.324 24h11.494v-9.294H9.689v-3.621h3.129V8.41c0-3.099 1.894-4.785 4.659-4.785 1.325 0 2.464.097 2.796.141v3.24h-1.921c-1.5 0-1.792.721-1.792 1.771v2.311h3.584l-.465 3.63H16.56V24h6.115c.733 0 1.325-.592 1.325-1.324V1.324C24 .593 23.408 0 22.676 0" />
+                  </svg>
+                </a>
+                <a href="#">
+                  <svg aria-labelledby="simpleicons-gmail-icon" role="img" viewBox="0 0 24 24"
+                    xmlns="http://www.w3.org/2000/svg">
+                    <title id="simpleicons-gmail-icon">Gmail icon</title>
+                    <path
+                      d="M24 4.5v15c0 .85-.65 1.5-1.5 1.5H21V7.387l-9 6.463-9-6.463V21H1.5C.649 21 0 20.35 0 19.5v-15c0-.425.162-.8.431-1.068C.7 3.16 1.076 3 1.5 3H2l10 7.25L22 3h.5c.425 0 .8.162 1.069.432.27.268.431.643.431 1.068z" />
+                  </svg>
+                </a>
+              </p>
+              <div class="footer-top-cap text-center">
+                <span><a href="#">inquiries@rnbrewery.com</a></span>
+                <span class="separator">|</span>
+                <span class="copyright">Rankdog & Nisbone Brewery &#169; 2021</span>
+                <span class="separator">|</span>
+                <span>Reston, VA, 20191</span>
+              </div>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+    <!-- Footer End-->
+  </footer>
+  <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js"
+    integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo"
+    crossorigin="anonymous"></script>
+  <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.7/umd/popper.min.js"
+    integrity="sha384-UO2eT0CpHqdSJQ6hJty5KVphtPhzWj9WO1clHTMGa3JDZwrnQq4sF86dIHNDz0W1"
+    crossorigin="anonymous"></script>
+  <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js"
+    integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM"
+    crossorigin="anonymous"></script>
+</body>
+
+</html>
