--- Code/J-Nistler/javascript/labs/lab01-JavascriptRedo/index.html
@@ -1,16 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Lab01 - Javascript Redo</title>
-    <script src="app.js"></script>
-</head>
-<body>
-    <h1>Lab01 - Javascript Redo</h1>    
-    <button id=rps onclick="rockPaperScissors()">Rock, Paper, Scissors</button>
-    <button id=rpg onclick="randPassGen()">Random Password Generator</button>
-    <button id=r13 onclick="rotCypher()">Rot13</button>
-</body>
-</html>
