--- Code/J-Nistler/javascript/labs/lab02-ToDoList/list.js
@@ -1,3 +1,0 @@
-let TODO = ['Go grocery shopping', 'Walk the dog', 'Do the dishes', 'Wash your brain']
-
-let DONE = ['Create Todo']
