--- Code/J-Nistler/javascript/labs/lab02-ToDoList/style.css
@@ -1,55 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-}
-
-.list_gen {
-  height: 500px;
-  overflow: auto;
-}
-
-.form-group {
-  display: flex;
-  justify-content: center;
-  margin-bottom: 30px;
-}
-
-.title {
-  display: flex;
-  justify-content: center;
-  margin-bottom: 30px;
-}
-
-.todo-check {
-  position: relative;
-}
-
-td { position: relative; }
-
-tr.strikeout td:before {
-  content: " ";
-  position: absolute;
-  top: 50%;
-  left: 0;
-  border-bottom: 1px solid #111;
-  width: 100%;
-}
-
-tr.strikeout td:after {
-  content: "\00B7";
-  font-size: 1px;
-}
-
-.strikeout + td {
-    visibility: hidden;
-}
-
-.badge, .rounded-pill, .bg-danger {
-  visibility: hidden;
-  position: relative;
-}
-
-.badge, .rounded-pill, .bg-danger:hover {
-  cursor: pointer;
-}
