+++ Code/J-Nistler/javascript/labs/lab04-ToDoList/list.js
@@ -1,0 +1,3 @@
+let TODO = ['Go grocery shopping', 'Walk the dog', 'Do the dishes', 'Wash your brain']
+
+let DONE = ['Create Todo']
