+++ Code/J-Nistler/python/lab09-Rot_Cypher.py
@@ -1,0 +1,72 @@
+#-----------------------------------------------#
+'''
+PDX Code Guild
+Class CodeBusters
+Lab 09 - Rot Cypher
+Jared Nistler
+'''
+#-----------------------------------------------#
+# Instructions
+#-----------------------------------------------#
+'''
+Write a program that prompts the user for a string, 
+and encodes it with ROT13. For each character, find 
+the corresponding character, add it to an output string. 
+Notice that there are 26 letters in the English language, 
+so encryption is the same as decryption.
+
+Version 2
+Allow the user to input the amount of rotation used in the 
+encryption. (ROTN)
+
+Version 3 (Optional)
+Add support for capital letters, numbers, and special characters. 
+These can be handled in two different ways:
+
+1) Capital letters can be rotated as well, numbers and special 
+characters can be put directly into the output (e.g. "hello world!" 
+becomes "uryyb jbeyq!").
+
+2) Instead of using an alphabet of just letters, include numbers, 
+spaces, and special characters as well.
+
+
+'''
+#-----------------------------------------------#
+# Developer Notes
+#-----------------------------------------------#
+'''
+Version 1: 
+Version 2: 
+Version 3:
+
+Future Improvements:
+
+
+'''
+
+#-----------------------------------------------#
+# Import Dependencies
+#-----------------------------------------------#
+
+
+#-----------------------------------------------#
+# Define Globals
+#-----------------------------------------------#
+
+
+#-----------------------------------------------#
+# Define Functions
+#-----------------------------------------------#
+
+
+
+#-----------------------------------------------#
+# User Input
+#-----------------------------------------------#
+
+
+#-----------------------------------------------#
+# Print Advice
+#-----------------------------------------------#
+
