+++ Code/J-Nistler/python/lab10-Unit_Converter.py
@@ -1,0 +1,84 @@
+#-----------------------------------------------#
+'''
+PDX Code Guild
+Class CodeBusters
+Lab 10 - Unit Converter
+Jared Nistler
+'''
+#-----------------------------------------------#
+# Instructions
+#-----------------------------------------------#
+'''
+Ask the user for the number of feet, and print out 
+the equivalent distance in meters. Hint: 1 ft is 0.3048 m. 
+So we can get the output in meters by multiplying the 
+input distance by 0.3048. Below is some sample input/output.
+
+Version 2
+Allow the user to also enter the units. Then depending on the 
+units, convert the distance into meters. The units we'll allow 
+are feet, miles, meters, and kilometers.
+
+Version 3
+Add support for yards, and inches.
+
+Version 4
+Now we'll ask the user for the distance, the starting units, 
+and the units to convert to.
+
+You can think of the values for the conversions as elements 
+in a matrix, where the rows will be the units you're converting 
+from, and the columns will be the units you're converting to. 
+Along the horizontal, the values will be 1 (1 meter is 1 meter, 
+1 foot is 1 foot, etc).
+
+But instead of filling out that matrix, and checking for each 
+pair of units (if from_units == 'mi' and to_units == 'km'), we 
+can just convert any unit to meters, then convert the distance 
+in meters to any other unit.
+
+Furthermore you can convert them from meters by dividing a 
+distance (in meters) by those same values used above. So first 
+convert from the input units to meters, then convert from meters 
+to the output units.
+
+'''
+#-----------------------------------------------#
+# Developer Notes
+#-----------------------------------------------#
+'''
+Version 1: 
+Version 2: 
+Version 3:
+Version 4:
+
+Future Improvements:
+
+
+'''
+
+#-----------------------------------------------#
+# Import Dependencies
+#-----------------------------------------------#
+
+
+#-----------------------------------------------#
+# Define Globals
+#-----------------------------------------------#
+
+
+#-----------------------------------------------#
+# Define Functions
+#-----------------------------------------------#
+
+
+
+#-----------------------------------------------#
+# User Input
+#-----------------------------------------------#
+
+
+#-----------------------------------------------#
+# Print Advice
+#-----------------------------------------------#
+
